fix: 빈 매장 keyword 변경 처리

+++ src/main/java/bdbe/bdbd/service/carwash/CarwashService.java
@@ -331,7 +331,9 @@ public class CarwashService {
         List<Long> keywordsToDelete = existingKeywordIds.stream()
                 .filter(id -> !newKeywordIds.contains(id))
                 .collect(Collectors.toList());
-        carwashKeywordJPARepository.deleteByCarwashIdAndKeywordIds(carwashId, keywordsToDelete);
+        if (!keywordsToDelete.isEmpty()) {
+            carwashKeywordJPARepository.deleteByCarwashIdAndKeywordIds(carwashId, keywordsToDelete);
+        }
 
         List<Long> keywordsToAdd = newKeywordIds.stream()
                 .filter(id -> !existingKeywordIds.contains(id))
